update dependencies; separate test cases; add waits; ignore messages due to timeout issues
