+++ chat-app/package.json
@@ -12,7 +12,7 @@
   "author": "fluvio <admin@fluvio.io> (fluvio.io)",
   "license": "Apache-2.0",
   "dependencies": {
-    "@fluvio/client": "^5.0.0-alpha.9",
+    "@fluvio/client": "^5.0.0-alpha.10",
     "chat-client": "file:chat-client",
     "chat-server": "file:chat-server"
   },
